requirements.txt
dj-rest-auth
django
django-allauth
djangorestframework
djangorestframework-simplejwt
django-environ